pyproject.toml
[tool.poetry]
name = "agenticdb"
version = "0.1.0"
description = ""
authors = ["BenderScript <6779302+BenderScript@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
langchain-chroma = "^0.1.2"
langchain-text-splitters = "^0.2.2"
langchain-community = "^0.2.7"
langchain-openai = "^0.1.16"
python-dotenv = "^1.0.1"
langserve = {extras = ["all"], version = "^0.2.2"}
langchain-core = "^0.2.18"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
